Replace `TesseractVariableName` with `TesseractOption` enum to handle Tesseract variables. This eliminates the need to add a property to `SwiftyTesseract` any time an addition to support a Tesseract addition is needed/requested.
Conform TesseractOption to Hashable and provide a custom implementation. This prevents allowing multiple variables of the
same name occupying a Set, even though they may have different values contained in them.

Create `InternalTesseractVariable` enum. This is to handle cases where a somewhat estoeric Tesseract variable
is needed to be set to make another one work. The one and only current case for this is `oldCharacterHeight`.
In SwiftyTesseract 2.x, for `minimumCharacterHeight` to work after setting a new value, `oldCharacterHeight` had to
be set to 1. This was hidden from the user and done in the didSet implementation of `minimumCharacterHeight` when
it was a property on SwiftyTesseract. Since the properties have been replaced in favor of the TesseractOption enum,
it posed an issue exposing `oldCharacterHeight` as a case because it feels clunky, confusing, and would be a burden on
the user of the library.

Create TesseractEnvironment enum. This really is soley for the purpose of setting the TESSDATA_PREFIX environment variable
for Tesseract to access. It didnt feel right to include it alongside oldCharacterHeight, and I wanted to avoid the ability
of being able to set oldCharacterHeight as an environment variable.